+++ server/src/main/java/com/codestates/julsinsa/global/dto/PageInfo.java
@@ -1,4 +1,4 @@
-package com.codestates.julsinsa.dto;
+package com.codestates.julsinsa.global.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
