+++ server/src/main/java/com/codestates/julsinsa/global/exception/BusinessLogicException.java
@@ -1,4 +1,4 @@
-package com.codestates.julsinsa.exception;
+package com.codestates.julsinsa.global.exception;
 
 import lombok.Getter;
