+++ server/src/main/java/com/codestates/julsinsa/member/service/MemberService.java
@@ -3,17 +3,14 @@ package com.codestates.julsinsa.member.service;
 import com.codestates.julsinsa.auth.dto.LoginDto;
 import com.codestates.julsinsa.auth.jwt.JwtTokenizer;
 import com.codestates.julsinsa.auth.utills.CustomAuthorityUtils;
-import com.codestates.julsinsa.exception.BusinessLogicException;
-import com.codestates.julsinsa.exception.ExceptionCode;
+import com.codestates.julsinsa.global.exception.BusinessLogicException;
+import com.codestates.julsinsa.global.exception.ExceptionCode;
 import com.codestates.julsinsa.helper.email.HtmlEmailSendable;
 import com.codestates.julsinsa.item.dto.ItemDto;
-import com.codestates.julsinsa.item.entity.Favorite;
 import com.codestates.julsinsa.helper.event.MemberRegistrationApplicationEvent;
-import com.codestates.julsinsa.item.entity.Item;
 import com.codestates.julsinsa.item.repository.FavoriteRepository;
 import com.codestates.julsinsa.member.dto.EmailRequest;
 import com.codestates.julsinsa.member.dto.FindDto;
-import com.codestates.julsinsa.member.dto.MemberDto;
 import com.codestates.julsinsa.member.entity.Member;
 import com.codestates.julsinsa.member.repository.MemberRepository;
 import io.jsonwebtoken.Claims;
@@ -23,7 +20,6 @@ import io.jsonwebtoken.security.SignatureException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.core.context.SecurityContextHolder;
