+++ server/src/main/java/com/codestates/julsinsa/order/service/OrderService.java
@@ -1,7 +1,7 @@
 package com.codestates.julsinsa.order.service;
 
-import com.codestates.julsinsa.exception.BusinessLogicException;
-import com.codestates.julsinsa.exception.ExceptionCode;
+import com.codestates.julsinsa.global.exception.BusinessLogicException;
+import com.codestates.julsinsa.global.exception.ExceptionCode;
 import com.codestates.julsinsa.member.entity.Member;
 import com.codestates.julsinsa.member.repository.MemberRepository;
 import com.codestates.julsinsa.order.entity.Order;
