+++ server/src/main/java/com/codestates/julsinsa/review/entity/Review.java
@@ -1,10 +1,9 @@
 package com.codestates.julsinsa.review.entity;
 
-import com.codestates.julsinsa.audit.Auditable;
+import com.codestates.julsinsa.global.audit.Auditable;
 import com.codestates.julsinsa.image.entity.ReviewImage;
 import com.codestates.julsinsa.item.entity.Item;
 import com.codestates.julsinsa.member.entity.Member;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
