+++ server/src/main/resources/application.yml
@@ -44,18 +44,9 @@ spring:
             scope:
               - email
               - public_profile
-#          naver:
-#            client-id: ${N_CLIENT_ID}
-#            client-secret: ${N_CLIENT_SECRET}
-#            scope:
-#              - name
-#              - email
-#            redirect-uri: http://localhost:8080/login/oauth2/code/naver
-#            authorization-grant-type: authorization_code
-#            client-name: Naver
           kakao:
-            client-id: 7c67eb4ba631e8116081b968377e35d6
-            client-secret: NWGmTyHL3wzQB2SdDDYAAeJdEmtom3Lw
+            client-id: ${C_CLIENT_ID}
+            client-secret: ${C_CLIENT_SECRET}
             redirect-uri: https://9b00-124-111-225-247.ngrok-free.app/login/oauth2/code/kakao
             authorization-grant-type: authorization_code
             client-authentication-method: POST
@@ -66,14 +57,6 @@ spring:
             token-uri: https://kauth.kakao.com/oauth/token
             user-info-uri: https://kapi.kakao.com/v2/user/me
             user-name-attribute: id
-#            naver:
-#              authorization-uri: https://nid.naver.com/oauth2.0/authorize
-#              token-uri: https://nid.naver.com/oauth2.0/token
-#              user-info-uri: https://openapi.naver.com/v1/nid/me
-#              user-name-attribute: response
-
-
-
 
 logging:
   level:
@@ -109,7 +92,6 @@ server:
   servlet:
     encoding:
       force-response: true
-
 
 jwt:
   key: ${JWT_SECRET_KEY}
